Operations with money are now in BigDecimal

+++ src/main/java/by/skoriyVladislav/dal/user_dao/UserDAO.java
@@ -2,6 +2,7 @@ package by.skoriyVladislav.dal.user_dao;
 
 import by.skoriyVladislav.entity.user.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserDAO {
@@ -9,8 +10,8 @@ public interface UserDAO {
     User createUser(String login, String pass);
     User createUser(String login);
     boolean registerUser(String login, String password, String name,
-                         String surname, String role, double balance, String email);
+                         String surname, String role, BigDecimal balance, String email);
     boolean loginInDataBase(String login);
-    boolean checkBalanceForBet(String login, double size);
-    boolean transaktion(User user, double size);
+    boolean checkBalanceForBet(String login, BigDecimal size);
+    boolean transaktion(User user, BigDecimal size);
 }
